add webserver to app
The objective of this project is to provide an All-in-one solution
to encrypt and server encrypted files.

This commit adds a webserver running at a random port and provides
information as the status and port of the server to the App, and
also provides the option to the app control the server status,
starting or stopping it.